Update docker to use new entrypoint style
Since the input file is no longer specified first, this allows us to add
arguments without having to overwrite the entrypoint.  This does mean that
"simple" usage now requires the `-` parameter:

    docker run -i regula:latest - <some_file.cfn

I think this is outweighed by the advantages: this makes it easier to pass in
other arguments, without having to re-specify `/opt/regula/`.  For example:

    docker run --rm \
        --volume [HCL_DIRECTORY]:/workspace \
        fugue/regula \
        /workspace

Where we needed the following monstrosity before:

    docker run --rm \
        --entrypoint regula \
        --volume [HCL_DIRECTORY]:/workspace \
        fugue/regula \
        /workspace \
        /opt/regula