[RM-5916] Add link to rule documentation in output report
This is currently a bit hackish; I just added code to check the ID at
the end of the process and inject the link if it's something we have
remediation documentation for.  Longer term it would probably make
sense to move the link into the __rego_metadoc__ blocks in the rego
files, but there is some infrastructure stuff which needs to happen
first.