[RM-5934] Remove buckets_with_valid_policies (#186)
`policy_document_ref_or_json_string` is meant to decide whether or not a policy
document is valid.  However, there was a bug in this function, resulting in all
inputs being treated as valid if a single policy policy data document exists.

However, this is only used in `s3_library.buckets_with_valid_policies`, which is
a bit of an antipattern: it first checked which buckets only have valid (valid
as in well-formatted JSON) policies, and then uses these in whatever rule.

It is much better to just iterate over all buckets and policy pairs, and discard
only the invalid ones.  That way a single invalid policy doesn't cause our
system to skip the other policies.