Add compact output mode. (#195)
Instead of writing the full URL on a separate line, this version just
links the text of the ID to the correct URL.  Unfortunately not all
terminals support this capability; there is a slightly out-of-date
overview of which emulators currently support them at
<https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda>.
I am not aware of a way to detect support at runtime.

Additionally, if there is only one item in the stack, the stack trace
will just be smushed onto the previous line so each violation only
takes a single line to report.