Makefile: unconditionally regenerate links for remediation target
This also adds remediation as a dependency for the bump targets so it
is automatically run when creating a new release.  I also removed the
rego/remediation.yaml target, which wasn't really useful since the
file is included in git not regenerated every time.

I have intentionally not updated rego/remediation.yaml to make testing
easier.